requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
Markdown==3.3.6
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
zipp==3.7.0
